+++ src/ExcelHandler.ts
@@ -21,7 +21,7 @@ export class ExcelRegistry implements IDisposable {
 
     getExcel() : Excel.Application {
         if ( !this.excel || !this.excel.Workbooks){
-            this.logger("EditExcelPQM: Create new Excel application");
+            this.logger("Create new Excel application");
             this.excel = new ActiveXObject('Excel.Application');
             (this.excel as any).ShowStartupDialog = false;
             this.excel.Visible = true;
@@ -30,18 +30,18 @@ export class ExcelRegistry implements IDisposable {
     }
 
     dispose(): void {
-        if (this.isInitialized() && this.getExcel().Workbooks.Count == 0){
+        if (this.isInitialized() && this.getExcel().Workbooks.Count === 0){
             this.getExcel().Quit();
         }
     }
 
     __getWorkbookFromCache(filename : string) : Excel.Workbook | undefined {
         if (this.isInitialized()){
-            let excel = this.getExcel()
+            let excel = this.getExcel();
             for(let i=1; i<=excel.Workbooks.Count;i++){
                 let wb = excel.Workbooks.Item(i);
-                if (wb.FullName == filename){
-                    this.logger("EditExcelPQM: Workbook is already open. Retrieve it.")
+                if (wb.FullName === filename){
+                    this.logger("Workbook is already open. Retrieve it.");
                     return wb;
                 }
             }
@@ -52,21 +52,22 @@ export class ExcelRegistry implements IDisposable {
     open(filename : string) : Excel.Workbook {
         let wb : Excel.Workbook | undefined = this.__getWorkbookFromCache(filename);
         if (!wb){
-            this.logger("EditExcelPQM: Open workbook with Excel.")
+            this.logger("Open workbook with Excel.")
             wb = this.getExcel().Workbooks.Open(filename);
         }
         return wb;
     }
 
-    close(filename : string) : void {
+    close(filename : string, saveChanges : boolean) : void {
         let wb : Excel.Workbook | undefined = this.__getWorkbookFromCache(filename);
         if (wb) {
-            this.logger("Close workbook " + filename)
-            wb.Close(true);    
+            this.logger("Close workbook " + filename);
+            wb.Close(saveChanges);    
+            wb = undefined;
         }
 
-        if (this.excel && this.excel.Workbooks.Count && this.excel.Workbooks.Count == 0){
-            this.logger("EditExcelPQM: All workbooks are closed. Close Excel also.")
+        if (this.excel && this.excel.Workbooks.Count !== undefined && this.excel.Workbooks.Count === 0){
+            this.logger("All workbooks are closed. Close Excel also.");
             this.getExcel().DisplayAlerts = false;
             this.getExcel().Quit();
             this.excel = undefined;
@@ -87,7 +88,7 @@ export class PowerQueryMCodeReader implements IDisposable {
 
     constructor(fileName : string, excelRegistry : ExcelRegistry){
         if (! fs.existsSync(fileName)){
-            throw new Error("EditExcelPQM: File not found ${fileName}");
+            throw new Error("File not found ${fileName}");
         }
         this.excelRegistry = excelRegistry;
         this.queries = new Map();
@@ -100,7 +101,7 @@ export class PowerQueryMCodeReader implements IDisposable {
         } else if (fileName.toLowerCase().endsWith(".xlsx") || fileName.toLowerCase().endsWith(".xlsm")) {
             this.sourceType = SourceType.Excel;
         } else {
-            throw new Error("EditExcelPQM: Unsupported format ${fileName}")
+            throw new Error("Unsupported format ${fileName}");
         }
 
     }
@@ -114,7 +115,7 @@ export class PowerQueryMCodeReader implements IDisposable {
         let workbook : any = this.excelRegistry.open(this.excelFileName);
         let queries = workbook["Queries"];
         if (queries === undefined){
-            throw new Error("EditExcelPQM: Queries attribute of Excel Workbook is undefined. It could happen due to raise of Activation window on Excel startup. Check Excel and try again.");
+            throw new Error("Queries attribute of Excel Workbook is undefined. It could happen due to raise of Activation window on Excel startup. Check Excel and try again.");
         }
         for(let i=1; i<=queries.Count; i++){
             this.queries.set(queries.Item(i).Name, queries.Item(i).Formula);
@@ -155,7 +156,7 @@ export class PowerQueryMCodeReader implements IDisposable {
         let workbook : any = this.excelRegistry.open(this.excelFileName);
         let excelQueries = workbook["Queries"];
         if (excelQueries === undefined){
-            throw new Error("EditExcelPQM: Worbook.Queries is undefined. Unable to import. This could be due to Excel pop-up windows.");
+            throw new Error("Worbook.Queries is undefined. Unable to import. This could be due to Excel pop-up windows.");
         }
         // make a copy of a queries map
         let queriesCopy : Map<string, string> = new Map(this.queries);
@@ -168,13 +169,13 @@ export class PowerQueryMCodeReader implements IDisposable {
                 queriesCopy.delete(name);
             } else {
                 // if query exist in Excel, but missed in M file, then drop it from excel
-                excelQueries.Remove(i);
+                excelQueries.Item(i).Delete();
                 i--;
             }
         }
         // now I need to create new items for ones thar were not present in Excel
         for(let [name, formula] of queriesCopy){
-            excelQueries.Add(name, formula)
+            excelQueries.Add(name, formula);
         }
     }
 }