+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 # Change Log
 
-All notable changes to the "editexcelpwm" extension will be documented in this file.
-
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+All notable changes to the "EditExcelPQM" extension will be documented in this file.
 
 ## [Unreleased]
 
-- Initial release
+## [1.0.1] - 2020-08-08
+### Added
+- After you delete query from M file and export it to xlsx, a deleted query is dropped in xlsx
+- Context menue now has an option to close excel without saving
+### Changed
+- All items of context menue have EEPQM prefix
+
+## [1.0.0] - 2020-07-26
+### Added
+- Export all M queries from xlsx/xlsm file to *.m file
+- Import queries from *.m file to xlsx/xlsm
+- Edit M code in VSCode and run queries in Excel immediately
+- Create new queries in VSCode and upload them to Excel